[FIX] hr_holidays: Disable emails when importing from file
When using the "test import" feature to import leaves request
allocations, odoo fails with a 500 error and no error message.

The only difference with an actual import is that the test dry-runs by
roll-backing. The issue here are that mails are sent *after* a commit
is done hence crashes because they do not exist any more.

Fine-tuning of cf9344c0dc6107349b211667c7100c8befdec0d3 for
hr.leave.allocation

opw-2041872

closes odoo/odoo#35166

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>