[FIX] base, crm: placeholder for website
Before this commit, the placeholder for the website was incorrect, it
missed the protocol. This could lead to a confusion to the user who
could add a website without protocol. A website without protocol will be
open in the browser as a page of the same website and not as an
independent website.

Now, the placeholders have the correct protocol.

opw-2042816

closes odoo/odoo#35217

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>