[FIX] website_sale: add tests for previous fix in lower version
In saas-14, 44761b230b275 was recently added to fix the flow where an empty
pricelist code being submitted in the cart input should reset the pricelist.

But we waited 12.0 to add an easy test, as we have `MockRequest` to easily
mock and simulate frontend context to test controller in python tests, see
e8ef497a82d2.

Following #34711

closes odoo/odoo#35253

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>