[FIX] base: check group constraints when adding implied groups
Since commit 59336792ec2, the write of implied groups is done directly in SQL,
bypassing the ORM writes, and thus bypassing the python constraints checks.
In particular such a constraint was added in 7284132b7fe.
We add back an explicit check of this constraint;
however we rewrite it in SQL to be fast.

closes odoo/odoo#35411

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>