[FIX] stock: do not raise error
Although this makes sense to warn the user, a `UserError` stops the
execution of onchanges but do not prevent saving incorrect settings. A
non-blocking warning would be another solution, but one issue remains:
a pop-up is raised when the user opens the settings, even if they are
unrelated (e.g.: open Website settings, an error message implying
Inventory appears).

This reverts commit 2fbd8406de3c93bad0cef047eb0ad7a455377ab3.

closes odoo/odoo#35511

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>