[FIX] account: assert_balanced in multi-company/multi-currency
The decimal precision must not be dependent of the current company's decimal places. E.g:
Create a move Z in the company A where decimale precision is 5.
Switch into the company B having a decimale precision of 8.
Edit the move Z, an error can appear because the decimale precision is now 8 instead of 5.

closes odoo/odoo#35572

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>