[FIX] *: remove buggy calls to with_context
When .with_context() is called with a dictionary as 1st positional
argument, it will replace context (and not modify the referenced keys)
It may create bugs when losing the content of the context (e.g. remove
partner's language)

This is a partial merge of #36164 without the inventory part as
discussed.

Closes odoo/odoo#36164

closes odoo/odoo#36691

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>