[FIX] event: Allow to navigate from registration to event
This patch fixes a functional regression introduced in 94b45f266503fb44159dea48fca54e52047f81f7. Before that commit, users were able to navigate from a registration to the related event. Now it's impossible, and under some circumstances this confuses the user (i.e. when he comes to the form from an URL that another user shared to him, and there are no breadcrumbs to follow).

Now, users will be able to click in the event and go to its form.

closes odoo/odoo#36414

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>