[FIX] account: chart of account installation cannot find a transfer account
Before this commit:

For a couple of clients, the installation of a chart of accounts crashes
because a transfer account is not found. This account should have been created
automatically, but hasn't, or has been deleted.

After this commit:

If no transfer account is found, no account is linked to the "Transfer account"
field of the company, but the installation doesn't crash.

closes odoo/odoo#36512

Opw: 2066557
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>