[FIX] web: add kw on web controller route
Before this revision, it wasn't possible to access attachments
stored in the filestore added in a mass mailing campaign. Indeed, the
URL uses the route `/web/image` (even for non-image content) but the
link tracker adds the following args: `utm_source` and `utm_medium`.

Those aren't recognized by the system and trigger an Internal Server
Error.

Adding the kw is allowing the public attachment to be seen without being
connected to the instance, which should be the case on mass mailing
campaigns. There is no need to do anything with those args here, only
not refusing them.

opw-2060831

closes odoo/odoo#36815

Signed-off-by: Christophe Simonis <chs@odoo.com>