[FIX] web: pass the context when executing an action
In case you have a button of type action, the controller is doing an rpc
on the model using the right method, but in this rpc, the context is not
passed.

This creates a translation problem (nothing is translated) since the
language is not defined in the context.

Task Id 2062419
Closes #36189

closes odoo/odoo#36906

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>