[FIX] base_import: clear cache after a dry run
Test import of a file with xml ids.
At the first dry run, the xml id X is associated with id R for each such record.
If some records refer to X via a relational field, then in SQL it reduces
to a query using R as id; but R does not exist since it was a dry run.
As a result, subsequent runs fail.

The cache should be cleared after a dry run, since the xml ids are not reliable.

opw 2068446

closes odoo/odoo#37370

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>