[FIX] rating: fix image computation and display
Currently if we try to display image of a rating having a different value of
0/1/5/10 there is an infinite loading. Indeed images for rating are limited
to those values.

In this commit we fix that by computing a rounded value to find the closest
image.

Text display in form view is also centered for better wow onboarding back 2
basics demo usability fiximp effect.

This is a cherry-pick of f518a7dc637122ad320b49bfb5e996fb71961d95
which is a bugfix commited in 13.0.
Even if the odds it happens to have another value than
0/1/5/10 are smaller, the issue can still happen,
for instance if the rating was set manually
e.g.
```python
rating = env["rating.rating"].create({'rating': 3, 'res_id': 1, 'res_model_id': 199})
rating.rating_image
```

Upgrade request 53535
M1712236092310

closes odoo/odoo#57304

X-original-commit: f643a32c943ed2a1fd5e11be26308a6343585802
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>