[FIX] account: method call consistency
Following commit 44f6e2cf166a1581451fe, we explicitly use the `product`
parameter to check the object type. This enlightened the fact that the
method `compute_all` might be called with `product` being an integer
instead of a recordset.

This was not noticed before because `product` is only used in
`account_tax_python`.

opw-2076993

closes odoo/odoo#37558

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>