[FIX] report_stock_forecast: wrong date type
Before this commit, the SQL view returned the date field as a datetime
kind, which wasn't compatible with the python date field. This arises an
error if we create a tree/list view of the report stock forecast.

Now, the SQL view returns the date field as a text kind, which is
compatible with the python field.

opw-2081255

closes odoo/odoo#38395

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>