[FIX] base, website: debug assets in multi-website/multi-company
Fine-tunning of 67828f7ce3183c0f2b921881d4bd003866533c49

Before this commit, the assets were not found when the user is connected
in mode debug assets in a multi-website and multi-company. This occurs
because the attachments (assets) were created only for the first website
accessed, and cannot be loaded when trying to access the second website.

Now, the attachments are created for each website accessed and can be
loaded withtout problems.

closes odoo/odoo#38942

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>