[FIX] web_editor: prevent creating editors for the same element twice
It was technically possible that two editors for the same element are
created in the same JS stack execution... and since 13.0 and the
use of native promises, it happens everytime a block is dropped in the
page.

Note: this commit fixes the problem with minimal code but the logic
should be improved in master.

closes odoo/odoo#39593

X-original-commit: 54d032bb66013b32fb4def2216f77f7890ce31df
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>