[FIX] l10n_in: fix empty HSN on invoice report & missing HSN column
Reproduce this bug:
 - Create 2 companies: Indian and Belgian for example
 - Install accounting & l10n_in
 - Create an invoice & print it

The `<td>` dedicated to the HSN is always present so there is a
cosmetic issue for the not indian companies: the table is a *little*
bit too large. This issue is caused by a missing conditional statement
on the `<td>` who is present on the HSN `<th>`.

OPW-2090663

closes odoo/odoo#39815

Signed-off-by: Jason Van Malder <jasonvanmalder@users.noreply.github.com>