[FIX] hr_holidays: removing widget handle from list view
In the Time Off/Leaves app, there are leaves types.
Go on the list view, there is the handle widget on each line.

We don't want this widget because it is misleading. The leaves types
have a specific order algorithm and cannot be reordered.

Fine tuning of 160a08c2

OPW-2089593

closes odoo/odoo#39865

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>