[FIX] web_editor: remove destroyed editors from memory
Commit https://github.com/odoo/odoo/commit/2914d328cd9ba72457003b9d2968987b8403f523
introduced a method to destroy all current editors to refresh them when
necessary... but forgot to empty the array that contains them.

In a 12.0 without custo, this is only slowing down the editor but in
master (or with custo) this makes the editor crash.

closes odoo/odoo#40335

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>