[FIX] website_blog: fix link addable on blog cover
A customer reported an issue when he tried to add a link on the
blog covers of his blog. The link is displayed as text and it's added
to all the covers.

![links on all covers](i.imgur.com/fuFu6op.jpg)

According to SBU, the user should not do this.

This commit prevents the user from editing a blog cover.

OPW-2119924

closes odoo/odoo#40695

X-original-commit: 386ead3290af0df9b41d2af2a858001eebb7d2cc
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>