[IMP] stock: track changes on type
Changing a product type from consumable to storable and and vice versa can
make the quantity in stock confusing. As only storable product update the
stock quants, the amount on stock moves could be diffenrent that the one
on stock quant if the product type has been changed in the past.

This commit will make those changes saved in the chatter history in order
to easily track inconsistencies

closes odoo/odoo#40933

Opw: 2125124
X-original-commit: 7f07d07c3559cb2885a192fa2210d9673612c246
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>