[FIX] mrp: decrease produced quantity
- Create a product A with a BOM
- Create a MO for 3000 units of A
- Product 3000 units
- Unlock the MO and change the Consumed Materials and Finished Products
  quantity to less than 3000, e.g. 2900
- Update the Quantity To Produce to 2900

The Finished Products move lines are deleted, but the Consumed Materials
are left unchanged.

This is because decreasing the quantity triggers a `_do_unreserve` which
deletes the move lines:

https://github.com/odoo/odoo/blob/d71eaf16a6bd103f253cd7f371a75ab52275f0fb/addons/stock/models/stock_move.py#L399

The issue is similar to ed4c46004bdd9846, so we use the same solution:
unreserve, but keep track of the quantities.

opw-2151927

closes odoo/odoo#41933

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>