[FIX] stock: free_reservation performance
Validating a move line will remove the reservation on other move line
that will take quantity on the same quant.
The querry searching for those move lines needs the state of their
relative stock move to be neither 'done' nor 'cancel'. This constraint
add a join in the SQL querry that slow largely the request

This commit removes the `move` state in the querry and replace it by the
the move line state to speed it up.
Thanks to
https://github.com/odoo/odoo/commit/e19584139e2b94543aac1913ec253001dd8b1df4
and
https://github.com/odoo/odoo/commit/fa579558ed53d818e2242551e5326178875cea31,
it's will give the same result as the
move line state is related stored of its relative move.

opw 2126433

closes odoo/odoo#42133

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>