[FIX] account: compute of amount residual in foreign currency
When having an invoice in foreign currency and a partial payment in
domestic currency (in a different date with a different exchange rate),
the computed amount due is incorrect. When computing the
amount_residual_currency the payment (in domestic currency) is converted
to the foreign currency using the exchange rate of the invoice date, and
not the one of the payment date (see 3a4137d38b7d0115228ae47e5dddef6bf9e3f3ae).

If the account.partial.reconcile line will have a currency_id and a
amount_currency the amount_residual_currency will not be converted but
simply added.

This behaviour was introduced with 0f0dd0189e9c25519f43198536208652388d0fff
and backported to version 11.0 with 3363d7235137f286458c734a1a0417beeb1f66bb.
In version 12.0 this behaviour was partially reverted in d7bec5d3ce7be8331cae04a0f56b48fa9deef207.

opw-2156001

closes odoo/odoo#42524

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>