[FIX] base: do not compute source with empty name
Before this commit, when creating a new Translated Term and selecting
the 'Model' type without entering name first, an error occured as
there was no name to be split (bool instead of str).

After this commit, skip the computation if the 'name' field is empty.

As the field is required, it won't be possible to save a record
without entering a name but if the type was entered before the name,
an error occured in the onchange.

closes odoo/odoo#42442

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>