[FIX] sale: use correct currency for reverse pricelist price computation
When the discount of a sales order is computed, based on the pricelist,
the pricelist price computation is reversed, to find the base price on which
the discount was applied.

The currency considered as the base price currency was potentially wrong in
some cases (multi-currency environment, different currency between cost and sales prices,
...).

This commit ensures the right currency is used.

X-original-commit: 3f3db89ece3f8df9bfa58e278843d07a84c66741
Co-authored-by: Victor Feyens <vfe@odoo.com>