[FIX] sale_management: No discount displayed in option products
[FIRST ISSUE]

Steps to reproduce the bug:

- Create a pricelist PL with 50% reduction on all products.
- Create a SO with PL
- Add an optional product P

Bug:

 The price is correct for P  but the discount displayed was 0%.

[SECOND ISSUE]

Steps to reproduce the bug:

- Add the option line to the order

Bug:

A new line was create for P in the SO with the correct price but the discount was not displayed

PS: The discount field on the model sale.order.option has been implemented to add a new discount
to the displayed price unit of the option. But it has not been implemented to show the discount computed with
the pricelist.

opw:2186470

closes odoo/odoo#44703

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>