[FIX] web_tour: rtl tooltips are aligned properly
Steps to reproduce:
- install sales
- install any rtl language and set your user language to it
- go to sales

this.info.position can have any of the 4 values 'bottom', 'top',
'right' and 'left'. It only needs to be reverted in rtl if it's left
or right.

Previous behavior:
tooltip is not positionned properly and covers the create button

Current behavior:
toolips align the way they are supposed to

opw-2193479

closes odoo/odoo#45951

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>