[FIX] fields: writing on many2many field drops inactive corecords
Use case: edit a group to remove some user.  If the superuser belongs to
that group, it is automatically dropped from the relation, because that
user is inactive.  The edition aimed at modifying the set of active
users of the group, but the field update actually modifies the set of
all users of the group.

Fix the method `Many2many.write` by considering the values that `read`
would return, and only modify those values.  With that fix, the update
only modifies the active users.  To update all users, one has to pass
`active_test=False` in context.

opw-2196358

closes odoo/odoo#46462

Signed-off-by: Raphael Collet (rco) <rco@openerp.com>
Co-authored-by: Nicolas Lempereur <nle@odoo.com>