[FIX] purchase_stock: deliver qui with different UoM
- Create a product A with UoM set to 'Units(s)'
- Create a product B with UoM set to 'Liter(s)'
- Create a BOM of type kit, 2 Liters of B to create 1 Unit of A
- Create a PO with A, validate
- Validate the corresponding picking

An error is raised since we try to compute the quantity received in
Units based on a quantity in Liters.

In case of a kit BOM, an override exists in `purchase_mrp` to compute
correctly the quantity. Therefore, we can skip the line.

Corresponding fix for SO: 3b2eee2fabaa363cdf5f9bad3547df8ef66f7d47

Closes #45871
opw-2205970

closes odoo/odoo#46687

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>