[FIX] product: apply fixed price digits in pricelist item form view
Since unit prices in sales order lines and list prices in the product
form view also use the same decimal precision (`Product Price`), it
makes sense to show all the precision digits when dealing with fixed
prices in a pricelist rule.

While the precision on the field's declaration was correct, its widget
used the decimal precision of the currency by default. This commit
forces the monetary widget to use the field's precision instead.

closes odoo/odoo#46705

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>