[FIX] l10n_ch: make QR-bill and ISR generation disjoint
QR-bills could only be generated for ISR-compliant invoices, which is wrong and would have caused issues in the future, as ISR will slowly disappear.

Furthermore, doing it this way avoids raising an error after commit https://github.com/odoo/odoo/commit/7497481605d08e33c66a100ca85ca50864fe9033 , when trying to print an ISR-compliant invoice if the module hadn't been updated (it used to raise because the QR-bill report didn't exist, and it tried generating it).

closes odoo/odoo#50744

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>