[FIX] website_sale: token checkout payment confirmation
Revision 6095d71 factorized some JS code from the ecommerce so that it
may be cleanly overriden in a 'fix' module required for a Stripe update.

Unfortunately, it broke the confirmation mechanism when paying with a
pre-recorded token in the eCommerce flow.

This commit reintroduces the required event handler so that token
payment confirmation works again.

opw-2283661