[FIX] event: duplicated cancel button
The 'Cancel Event' button appears twice on the form view of an event.

The issue was introduced by 59cc53e95be7674: by removing the `|`, a
`AND` expression is used to combine the `states` attributes with the
`invisible` `attrs`.

We combine the domain with the expected `|` operator to fix the issue
and avoid re-introducing the problem described in the aforementioned
commit.

opw-2296838

closes odoo/odoo#54773

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>