[FIX] mail: allow bad encoding on attachment content-id
When walking over a email.message.Message, if the `content-id` of an
attachment contains bad encoding, we might get a mail.header.Header:

https://docs.python.org/3/library/email.header.html

object instead of an expected `str`. A mail containing that would raise
an error an not appear in Odoo because code expected `str`.

Without the fix, the added test breaks with:

    File ".../mail_thread.py", line 1590, in _message_extract_payload
      inner_cid = part.get('content-id').strip('><')
    AttributeError: 'Header' object has no attribute 'strip'

opw-2274374
closes #56582

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>