[FIX] timesheet_grid: prevent timesheet duplication
- Create 2 projects PA & PB
- Create a task in project TA & TB
- Go to Timesheet > My Timesheet
- Create a timesheet on PA, for task TA
- In the timesheet list view, group by project
- Click on a timesheet => you access the form view
- Duplicate the task

It is possible to set the task TB and keep the project PA. Things get
inconsistent afterwards. Note that it is also possible to access the
timesheet from the kanban view and edit it from the form view.

We add a constrain to ensure the consistency of the selected task and
the project.

opw-2337086

closes odoo/odoo#58058

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>