[FIX] google_drive: fix traceback when add filter
Before this commit:

When trying to add a filter in the google drive template and try to open the
record, it shows traceback for some filter domain which contains
'uid',  'user',  'time', 'datetime', 'dateutil', etc.. and traceback occurring
because these are not evaluating.

After this commit:

This commit will prepare the context used when evaluating python code
and now google drive template will be workable for filters.

Task-2336393

closes odoo/odoo#58207

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>