[FIX] google_calendar: user-friendly error message on synchronisation fail
Give a chance to the user to solve the issue by himself
when the Google synchronization is failing,
by displaying the actual error returned by Google
regarding the error.

The case of the OPW behind this revision is
the synchronisation from odoo to google calendar
of an event with an attendee having an invalid email
address.

In such a case, Google returns the meaningful error:
"Invalid attendee email".

We better pass this information to the user,
along with the event causing the trouble,
so he gets a chance to solve the issue by himself.

I take into account the revision
4865e70fbcd3324a6ba6815a32a849ebddd036e1
Multiple attempts have been done by Julien
to solve this issue genericly,
by catching the error and raising an `UserError`
in the helper `_do_request` directly,
but this caused various issues for methods
expecting `_do_request` to return `HTTPError`,
which suddenly returned `UserError` instead.

I therefore chose to catch the error directly in `create_an_event`,
which is only called at one place,
and this place doesn't expect to catch `HTTPError` exceptions.
(no `try...except...`)

opw-2334442

closes odoo/odoo#58785

X-original-commit: 8e5fbca51771d1b19019df29a7bd145e26fa7676
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>