[IMP] sale_management: Non-deterministic elements in views
There are currently some elements in views that are not reachable on a
deterministic way when those views are inherited. That means, to reach
them, positional selectors need to be used (e.g. `last()`). That causes
inherited views to get broken when there are small changes in element
position.

A good example of this is [1]:

    <th class="text-right">Unit Price</th>

Which can't be filtered by:
- Class, because there are several `<th>` elements with the same class
- Text content, because when content is translated, selector will fail

To solve the above, this change provides IDs for several elements, to make
possible/easier to reach them on a safely manner.

[1] https://github.com/odoo/odoo/blob/f823dbc674b2/addons/sale_management/report/sale_report_templates.xml#L15

closes odoo/odoo#59453

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>