[FIX] sale: revert 4abd20c1a369e
Revert "[FIX] sale: correct _get_to_invoice_qty depends"
This reverts commit 4abd20c1a369eeb6f3899a38674c8ec761718fcb.

While the field is indeed in the list of records, it's a big change of
behaviour.
Retroactively changing the invoicable lines of sale.order that are
closed is dangerous and can lead on unexpected behaviours.
Performances can also be bad when changing the product configuration
on a database with a lot of sale orders.

Changing the product configuration should only apply to future orders,
not past.

closes odoo/odoo#61135

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>