[FIX] calendar: default view to the activity date when opening the calendar from an activity
Behavior prior to this commit:

When I edit an activity and use the "Open Calendar" button, the calendar
defaults to the current date, thus I don't see the activity I was
editing.  Even though this button is used to schedule a new activity
(which might be for the current date rather than the activity's), it is
surprising when opening the calendar FROM an activity.

Behavior after this commit:

When opening the calendar from an activity, the intial date is set to
the edited activity's date.

opw-2369347

closes odoo/odoo#61373

Signed-off-by: Nicolas Galler <ngaller@users.noreply.github.com>