[FIX] stock: apply push rule after transparent rule
Usecase to reproduce:
- Rule from WH/A -> WH/B as automatique no step added
- Rule from WH/B -> WH/C as transfer created

Do a move that goes to WH/A and confirm it. It's destination location
is correctly set to WH/B but the internal transfer from WH/B to WH/C
is not created.

It's due to the avoid loop location that returns always true due to
code removed in commit d136c56e10b0ae2429f88ccc2013e2d1b89ef460

closes odoo/odoo#61414

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>