[FIX] CVE-2018-15638, mail: use consistent placeholder
When generation the message. add the name of the channel with a
placeholder. In some languages like Japanese, the order of words is
different and adding the name at the end makes an incorrect sentence,
it was not possible to properly translate it.

Affects: Odoo 13.0 and earlier (Community and Enterprise Editions)
Severity :: High :: 7.1 :: CVSS:3.0/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:L/A:N
Cross-site scripting (XSS) issue in mail module in Odoo Community 13.0
and earlier and Odoo Enterprise 13.0 and earlier, allows remote attackers
to inject arbitrary web script in the browser of a victim via crafted
channel names.

https://github.com/odoo/odoo/issues/63703