[CHG] base_import: increase recursion limit to 3
As part of QoL PR #26267 the way the recursion limit was
managed (where it is checked) was changed, effectively decreasing the
recursion limit by 1 (to the originally expected limit of 2).

According to #29877 this is an inconvenient change / regression in
e.g. accounting context: if an object has journal entries, updating a
field of a journal item requires 3 levels of indirection (entries ->
items -> field). The same would occur if an object is e.g. linked to
multiple projects (projects -> tasks -> field) or SO (-> lines ->
field).

Therefore bump the recursion limit up to 3.

Closes #29877

closes odoo/odoo#61934

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>