[IMP] CVE-2020-29396, base: make datetime coherent with dateutil
As we do not need all the datetime module, we want to restrict the
functions (as we do for dateutil). Also centralise handling of this
and apply it by default, no reason for there to be divergences.

Task 2179629

Affects: Odoo 11.0 through 13.0 (Community and Enterprise Editions)
Severity :: Critical :: 9.9 :: CVSS:3.0/AV:N/AC:L/PR:L/UI:N/S:C/C:H/I:H/A:L
A sandboxing issue in Odoo Community 11.0 through 14.0 and Odoo Enterprise
11.0 through 14.0, when running with Python 3.6, allows remote authenticated
users to execute arbitrary code, leading to privilege escalation.

https://github.com/odoo/odoo/issues/63712

In this port, add the missing datetime import as it is used to prepopulate
the wrapped module cache.