Revert "[FIX] google_drive: fix traceback when add filter"
This reverts commit cc79992da089937aff216f4797385a040d96d071.

The fix is actually not solving all the issues, and also introduce
a behavior change. Actually, the expected result for filters like

<filter string="My Pipeline" name="my" domain="[('user_id', '=', uid)]"/>

will result into a evaluated filters, but invalid as the uid is set to
the person configuring the google configuration, not the user who will
use it.

As it is tricky (or even impossible) to solve the issue properly in all
the cases, we prefer to warn the user that the filter is invalid instead.

closes odoo/odoo#64079

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>