[REV] account: reverse entry when account move line contains 0 in credit or debit
This fix partly reverts the commit 25064f76afe14a38655167e417c93967d2277251
The latter is fully reverted in v13: 6a5271b7f4109a4f2a20378698fbe479d73ed9cb

When having an invoice with total equals to 0, it still adds a line to
the follow-up report.

To reproduce the error:
1. Go to Accounting > Customers > Invoices
2. Create an invoice without Payment Terms and product at 0$ to have a total of 0$
3. Go to Accounting > Customers > Follow-up Reports > [Select the partner]
4. The invoice is present with a total due of 0$

Err: the invoice shouldn't be in the follow-up report.

OPW-2419268

closes odoo/odoo#64214

Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>