[FIX] sale_purchase: Invalid context mapping
When canceling a purchase order generated from a SO, a warning activity
is generated on the SO to warn the salesman that the PO was cancelled.

But the activity message was incomplete, because the cancelled PO lines
were given as 'purchase_lines' to the rendering context, but expected
as 'purchase_order_lines' in the qweb template.

This commit ensures that the correct key is used and that the activity
message is clear and complete.

Fixes #67508

closes odoo/odoo#67531

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>