[FIX] Adding error message instead of traceback when cropping pictograms
Description of the issue/feature this PR addresses:
When trying to crop pictograms, a traceback was issued because they are icons, not images.

Current behavior before PR:
The user was able to invoke the cropping action on pictograms and received an internal error.

Desired behavior after PR is merged:
The user now receives an error message stating that pictograms cannot be cropped and no traceback.

OPW: 2497084

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/odoo#70142

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>