[FIX] *: update documentation links
Following the recent reorganisation of the documentation in 12.0+, the majority
of the documents have been moved and their old links are no longer valid.

Some redirection rules will soon be deployed, but those rules might be dropped in some
years and we want the links to still work, which is why we still replace the links to the
new ones.

closes odoo/odoo#70639

Related: odoo/enterprise#18259
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>