[FIX] payment_stripe_sca: reference not found on return from checkout
Commit d0fc42c wrongly assumed that, in Stripe's redirect flow, the
feedback data dict could only have a metadata key if it was manually
passed in the payment request dict. In fact, a metadata key is always
present and can be associated with an empty dict.

After this commit, we no longer overlook the transaction reference and,
instead, introspect the feedback data too if the reference was not found
in the metadata.

closes odoo/odoo#71933

Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>