[FIX] google_calendar: make commit after user sync
syncing calendar for a single user may take few minutes, which could easily lead
to cron timeout. To avoid starting the syncronization over and over again, we
should commit updates that are already done.

The module already has many commit() calls, but those are not always used and
hence we need to add a new one.

---

opw-2577018

closes odoo/odoo#73054

Signed-off-by: Ivan Yelizariev // IEL <yelizariev@users.noreply.github.com>