[FIX] stock, purchase_stock: add product to PO if orderpoint
Suppose a PO that has been generated thanks to reordering rule. When
changing the product of a line, if the user runs the scheduler, the old
product won't be added to the PO

To reproduce the error
1. Create a product P:
    - Product Type: Storable
    - Purchase: Add a vendor
    - Reordering rule:
        - Min = Max = 1
2. Run the scheduler
3. Update the generated PO:
    - Change the product of the line
4. Run the scheduler again
5. Open the PO

Error: The PO only contains one line (the one with the other product).
The PO should now contains a second line for P

OPW-2538707

closes odoo/odoo#75333

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>