[FIX] Commit transient model vacuuming directly
Before this commit:
If the "Base: Auto-vacuum internal data" is archived since a long time, auto-vacuuming models may timeout. As the commit was previously done at the end of the cron, the process wouldn't have been applied. As a consequence the Scheduled action will timeout every time

After this commit:
The commit is done right after each model vacuuming

OPW-2638706

closes odoo/odoo#77730

Signed-off-by: Raphael Collet (rco) <rco@openerp.com>